Terraform module to deploy cronjob-like tasks on ECS.
In a diverse environment, it's common to have the requirement to provide an way
to execute scheduled operation on your infrastructure. Those tasks was sometimes
deployed via the, awesome, cronjob service. However, with the arrival of the
containers environment, a fully containerised ecosystem, will want to also
execute these queries by running a container on specific times. And so,
leveraging all perks from container: easily packing, versioning, deployments.

This module provides a solution that uses AWS EventBridge time-based event to
launch an ECS tasks on Fargate. Allowing users to deploy scheduled tasks. So,
we can "bring cron to AWS ECS ;)".