ecs: Fix missing of `awslogs-region` config for ECS task definition (#5)
When running a Fargate deployment, we need to specify the region of the
AWS CloudWatch Logs, which needs to be added as part of the
definition.LogConfiguration attribute.

We could add a new variable to tell us which AWS region to deploy the
changes. However, the current module, deploys all its resource on the
default AWS region from the user environment. Thus, let's use a
datasource to get the current default region. This will ensure the
configuration is comply with all the other resource creation.

Fixes #2